Moved symbols with double-underscore into internal namespace as they are meant to be private
